curPlayer 削除。

+++ app/src/main/kotlin/jp/ddo/masm11/contextplayer/player/Player.kt
@@ -318,11 +318,11 @@ class Player : Runnable {
 	}
     }
     
-    val currentPosition : Int
+    val currentPosition : Long
     get() {
 	val plr = curPlayer
 	if (plr != null)
-	    return plr.currentPosition
+	    return plr.currentPosition.toLong()
 	return 0
     }
